% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informationCriteria.R
\name{AIC}
\alias{AIC}
\alias{BIC}
\alias{loglik}
\title{Information Criteria}
\usage{
AIC(prediction, reference, model)

BIC(prediction, reference, model)

loglik(prediction, reference)
}
\arguments{
\item{prediction}{a vector or ts object of predicted values}

\item{reference}{a vector or ts object (on the same time scale as prediction) containing ground truth values}

\item{model}{a SlidingWindowReg model}
}
\value{
AIC or BIC value
}
\description{
Computes Akaike's Information Criterion (AIC) and Bayesian Information Criterion (BIC).
}
\section{Functions}{
\itemize{
\item \code{AIC()}: Akaike's Information Criterion

\item \code{BIC()}: Bayesian Information Criterion

\item \code{loglik()}: log-likelihood function

}}
