% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informationCriteria.R
\name{AIC.SWR}
\alias{AIC.SWR}
\alias{BIC.SWR}
\title{Information Criteria}
\usage{
\method{AIC}{SWR}(object, ts_input, ts_output, ...)

\method{BIC}{SWR}(object, ts_input, ts_output, ...)
}
\arguments{
\item{object}{an \code{SWR} model object created using \link{trainSWR}}

\item{ts_input}{a vector or ts object containing the input time series}

\item{ts_output}{a vector or ts object (on the same time scale as ts_input) containing the target time series}

\item{...}{currently unused}
}
\value{
AIC or BIC value
}
\description{
Computes Akaike's Information Criterion (AIC) and Bayesian Information Criterion (BIC).
}
\section{Functions}{
\itemize{
\item \code{AIC(SWR)}: Akaike's Information Criterion

\item \code{BIC(SWR)}: Bayesian Information Criterion

}}
