% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{eval_all}
\alias{eval_all}
\alias{rmse}
\alias{nrmse}
\alias{r2}
\alias{nse}
\alias{kge}
\title{Compute all available model metrics}
\usage{
eval_all(prediction, reference)

rmse(prediction, reference)

nrmse(prediction, reference)

r2(prediction, reference)

nse(prediction, reference)

kge(prediction, reference)
}
\arguments{
\item{prediction}{a vector or ts object of predicted values}

\item{reference}{a vector or ts object (on the same time scale as prediction) containing ground truth values}
}
\description{
computes all evaluation metrics given a prediction and a reference vector. The functionality includes general regression metrics and hydrological metrics (wrapper to \insertCite{bigiarini2020hydroGOF}{SlidingWindowReg}):
\itemize{
\item Root Mean Square Error (RMSE)
\item Normalized Root Mean Square Error (NRMSE)
\item Coefficient of Determination (R2)
\item Nash-Sutcliffe Efficiency \insertCite{nash1970nsc}{SlidingWindowReg}
\item Kling-Gupta Efficiency \insertCite{gupta2009kge}{SlidingWindowReg}
}
}
\section{Functions}{
\itemize{
\item \code{rmse()}: Root Mean Squared Error

\item \code{nrmse()}: Normalized Root Mean Squared Error

\item \code{r2()}: Coefficient of Determination

\item \code{nse()}: Nash-Sutcliffe Efficiency

\item \code{kge()}: Kling-Gupta Efficiency

}}
\references{
\insertAllCited{}
}
\seealso{
\link[hydroGOF]{hydroGOF}
}
