% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKernel.R
\name{get_kernel}
\alias{get_kernel}
\alias{plot_kernel}
\title{Return and plot accumulated kernel}
\usage{
get_kernel(param, mix = NULL, type = "single", weighted = TRUE)

plot_kernel(
  list = NULL,
  param = NULL,
  mix = NULL,
  type = "single",
  weighted = TRUE,
  xlim = NULL,
  include_text = TRUE
)
}
\arguments{
\item{param}{a matrix with 2 columns representing one window per row. The first column contains location parameters delta, the second column contains the standard deviation sigma.}

\item{mix}{a vector of mixing parameters (beta)}

\item{type}{a String indicating whether (a) all single windows (option: "single"), or (b) the combined kernel (option "combined")}

\item{weighted}{if TRUE, windows are weighted with mix parameters; default: TRUE}

\item{list}{a list containing multiple trained SlidingWindowReg models}

\item{xlim}{vector with lower / upper bound of the x axis to print the kernel}

\item{include_text}{if TRUE, plots will be annotated}
}
\description{
Return and plot the accumulated kernel, consisting of one or multiple windows
}
\examples{
param <- cbind(
  delta = c(0, 10),
  sigma = c(2, 3))
mix <- rep(1, ncol(param))
get_kernel(param = param, mix = mix)
plot_kernel(param = param, mix = mix)
}
