% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{window_overlap}
\alias{window_overlap}
\alias{overlap}
\title{Kernel overlap}
\usage{
window_overlap(param_prediction, param_reference)

overlap(
  param_prediction,
  param_reference,
  mix_prediction = NULL,
  mix_reference = NULL
)
}
\arguments{
\item{param_prediction}{predicted window parameter matrix (2 columns)}

\item{param_reference}{ground truth window parameter matrix (2 columns)}

\item{mix_prediction}{vector of predicted regression parameters}

\item{mix_reference}{vector of ground truth regression parameters}
}
\value{
a vector of overlap values for each ground truth window

a scalar indicating the overlap between predicted and ground truth kernel. Value between 0 (minimum overlap) and 1 (complete overlap).
}
\description{
Computes the kernel overlap as evaluation metric between a predicted kernel and the ground truth kernel
}
\section{Functions}{
\itemize{
\item \code{window_overlap()}: Computes the kernel overlap between each predicted window and each ground truth window separately

}}
