% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrediction.R
\name{plot_prediction}
\alias{plot_prediction}
\title{Plot prediction}
\usage{
plot_prediction(
  prediction,
  reference,
  input_ts = NULL,
  scale = "year",
  years = NULL
)
}
\arguments{
\item{prediction}{a vector or ts object of predicted values}

\item{reference}{a vector or ts object (on the same time scale as prediction) containing ground truth values}

\item{input_ts}{a vector or ts object of input data; if NULL, no input time series  will be plotted}

\item{scale}{a String indicating which scale should be used for plotting; options are "year", "month" or "week"; default is "year"}

\item{years}{a vector of years to be included in the plot; if NULL, the first 3 and last 3 years will be used}
}
\description{
plot the prediction and the reference
}
\examples{
param <- cbind(
  delta = c(0, 10),
  sigma = c(2, 3))
mixture <- rep(1, ncol(param)) # VERSION WITHOUT INTERCEPT: ncol(param) + 1
pred <- predict(sampleWatershed$rain, mixture, param)
plot_prediction(pred, sampleWatershed$gauge, sampleWatershed$rain)
}
