% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict target variable}
\usage{
predict(ts, mix, param, log = FALSE)
}
\arguments{
\item{ts}{a vector or ts object}

\item{mix}{a vector of mixing parameters (beta)}

\item{param}{a matrix with 2 columns representing one window per row. The first column contains location parameters delta, the second column contains the standard deviation sigma.}

\item{log}{whether a log-linear model should be used}
}
\description{
predicts the target variable given a time series of inputs, and trained parameters
}
