% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.SWR}
\alias{predict.SWR}
\title{Predict target variable}
\usage{
\method{predict}{SWR}(object, newdata, ...)
}
\arguments{
\item{object}{an \code{SWR} model object created using \link{trainSWR}}

\item{newdata}{a vector or ts object of new model inputs to predict}

\item{...}{currently unused}
}
\description{
Predicts the output time series given an input time series using a trained \code{SWR} model object.
}
