% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleWatershed2.R
\docType{data}
\name{sampleWatershed2}
\alias{sampleWatershed2}
\title{Sample watershed 2}
\format{
A data frame with 14153 rows and 3 columns:
\describe{
\item{date}{date variable (daily measurements) between January 01, 1980 and September 30, 2018}
\item{rain}{daily cumulated rainfall}
\item{gauge}{daily cumulated, gauged watershed runoff}
}
}
\source{
processed by HydroGeoScience for Watershed Management Research group at the University of British Columbia, Vancouver, Canada
}
\usage{
sampleWatershed2
}
\description{
A dataset containing a time series of daily rainfall and measured water runoff in a gauged watershed over 39 years
}
\keyword{datasets}
