% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlidingWindowReg-package.R
\docType{package}
\name{SlidingWindowReg-package}
\alias{SlidingWindowReg}
\alias{SlidingWindowReg-package}
\title{SlidingWindowReg: Gaussian Sliding Windows Regression}
\description{
A Sliding Windows Regression model with parameterized Gaussian kernels as weights for hydrological inference. The package is currently under development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sschrunner.github.io/SlidingWindowReg/}
}

}
\author{
\strong{Maintainer}: Stefan Schrunner \email{s.schrunner@fh-kaernten.at} (\href{https://orcid.org/0000-0003-1327-4855}{ORCID})

}
\keyword{internal}
