% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.SWR}
\alias{summary.SWR}
\title{Summarizes an `SWR` model}
\usage{
\method{summary}{SWR}(object, ...)
}
\arguments{
\item{object}{a `SWR` model object created using \link{train}}

\item{...}{currently unused}
}
\description{
A summary of an `SWR` model
}
