% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{encodeParam}
\alias{encodeParam}
\alias{decodeParam}
\alias{error}
\alias{train_inc}
\alias{train}
\title{Train model}
\usage{
encodeParam(mix, param)

decodeParam(x, intercept = FALSE)

error(ts_input, ts_output, param, mix, log)

train_inc(ts_input, ts_output, iter, log, param_selection = "best_bic")

train(
  ts_input,
  ts_output,
  iter = 10,
  runs = 10,
  log = FALSE,
  parallel = TRUE,
  return = "best",
  param_selection = "best_rmse"
)
}
\arguments{
\item{x}{encoded parameter vector}

\item{intercept}{if TRUE, a model with intercept is used}

\item{ts_input}{a vector or ts object containing the input time series}

\item{ts_output}{a vector or ts object (on the same time scale as ts_input) containing the target time series}

\item{log}{whether a log-linear model should be used}

\item{iter}{number of iterations (maximum number of windows)}

\item{param_selection}{either "max" (maximum number of windows), or "best_rmse", "best_aic", or "best_bic" to optimize RMSE, AIC, or BIC, respectively}

\item{runs}{number of independent model runs}

\item{parallel}{should the runs be computed in parallel? If FALSE, all runs are computed in serial. If TRUE, all runs are computed in parallel with a maximum number of cores. If a scalar is provided, the number of cores is set manually.}

\item{return}{either "best" (best model run is returned), or "all" (all model runs are returned)}
}
\description{
combines both training steps
}
\section{Functions}{
\itemize{
\item \code{encodeParam()}: Encode parameter vector

\item \code{decodeParam()}: Decode parameter vector

\item \code{error()}: Evaluate model error

\item \code{train_inc()}: Core training process

}}
\examples{
set.seed(42)
train(sampleWatershed$rain,
      sampleWatershed$gauge,
      iter = 2,
      runs = 1,
      parallel = FALSE)
}
