% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{trainSWR}
\alias{trainSWR}
\title{Train model}
\usage{
trainSWR(
  ts_input,
  ts_output,
  iter = 5,
  runs = 1,
  log = FALSE,
  parallel = FALSE,
  return = "best",
  param_selection = "best_rmse"
)
}
\arguments{
\item{ts_input}{a vector or ts object containing the input time series}

\item{ts_output}{a vector or ts object (on the same time scale as ts_input) containing the target time series}

\item{iter}{number of iterations (maximum number of windows)}

\item{runs}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} number of independent model runs; no longer supported due to deterministic window initialization}

\item{log}{whether a log-linear model should be used}

\item{parallel}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} should the runs be computed in parallel? If FALSE, all runs are computed in serial. If TRUE, all runs are computed in parallel with a maximum number of cores. If a scalar is provided, the number of cores is set manually. No longer supported for single-run models}

\item{return}{either "best" (best model run is returned), or "all" (all model runs are returned)}

\item{param_selection}{either "max" (maximum number of windows), or "best_rmse", "best_aic", or "best_bic" to optimize RMSE, AIC, or BIC, respectively}
}
\description{
Trains an \code{SWR} model based on input and target time series data.
}
\examples{
# train a model based on one year of observations
set.seed(42)
data(sampleWatershed)
mod <- trainSWR(sampleWatershed$rain[1:365],
                sampleWatershed$gauge[1:365],
                iter = 2,
                runs = 1,
                parallel = FALSE)
summary(mod)
}
