% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{eval_all}
\alias{eval_all}
\alias{rmse}
\alias{nrmse}
\alias{r2}
\alias{nse}
\alias{kge}
\title{Compute all available model metrics}
\usage{
eval_all(prediction, reference)

rmse(prediction, reference)

nrmse(prediction, reference)

r2(prediction, reference)

nse(prediction, reference)

kge(prediction, reference)
}
\arguments{
\item{prediction}{a vector or ts object of predicted values}

\item{reference}{a vector or ts object (on the same time scale as prediction) containing ground truth values}
}
\description{
computes all evaluation metrics given a prediction and a reference vector
}
\section{Functions}{
\itemize{
\item \code{rmse()}: Root Mean Squared Error

\item \code{nrmse()}: Normalized Root Mean Squared Error

\item \code{r2()}: Coefficient of Determination

\item \code{nse()}: Nash-Sutcliffe Efficiency

\item \code{kge()}: Kling-Gupta Efficiency

}}
\seealso{
\link[hydroGOF]{NSE}

\link[hydroGOF]{KGE}
}
