% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{convolve_window}
\alias{convolve_window}
\alias{build_gaussian_kernel}
\alias{predict}
\title{Predict target variable}
\usage{
convolve_window(ts, kernel)

build_gaussian_kernel(param)

predict(ts, mix, param, log = FALSE)
}
\arguments{
\item{ts}{a vector or ts object}

\item{kernel}{a window kernel}

\item{param}{a matrix with 2 columns representing one window per row. The first column contains location parameters delta, the second column contains the standard deviation sigma.}

\item{mix}{a vector of mixing parameters (beta)}

\item{log}{whether a log-linear model should be used}
}
\description{
predicts the target variable given a time series of inputs, and trained parameters
}
\section{Functions}{
\itemize{
\item \code{convolve_window()}: Convolve the time series with the window kernel

\item \code{build_gaussian_kernel()}: Build a Gaussian window kernel

}}
