% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.SWR}
\alias{predict.SWR}
\title{Predict target variable}
\usage{
\method{predict}{SWR}(object, newdata, ...)
}
\arguments{
\item{object}{an `SWR` model object created using \link{trainSWR}}

\item{newdata}{a vector or ts object of new model inputs to predict}

\item{...}{currently unused}
}
\description{
predicts the target variable given a time series of inputs, and trained parameters
}
