% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleWatershed.R
\docType{data}
\name{sampleWatershed}
\alias{sampleWatershed}
\title{Sample watershed}
\format{
A data frame with 14245 rows and 3 columns:
\describe{
  \item{date}{date}
  \item{rain}{daily rainfall}
  \item{gauge}{daily gauge value}
}
}
\source{
internal (UBC)
}
\usage{
sampleWatershed
}
\description{
A dataset containing a time series of daily rainfall and measured water flow in a gauged watershed over 39 years
}
\keyword{datasets}
