% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.SWR}
\alias{summary.SWR}
\alias{dim.SWR}
\title{Summarizes an `SWR` model}
\usage{
\method{summary}{SWR}(object, ...)

\method{dim}{SWR}(x)
}
\arguments{
\item{object}{an `SWR` model object created using \link{trainSWR}}

\item{...}{currently unused}

\item{x}{an `SWR` model object created using \link{trainSWR}}
}
\description{
A summary of an `SWR` model
}
\section{Functions}{
\itemize{
\item \code{summary(SWR)}: Summarizes the `SWR` model

\item \code{dim(SWR)}: Returns the number of windows of the `SWR` model

}}
