% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train model}
\usage{
train(
  ts_input,
  ts_output,
  iter = 10,
  runs = 10,
  log = FALSE,
  parallel = TRUE,
  return = "best",
  param_selection = "best_rmse"
)
}
\arguments{
\item{ts_input}{a vector or ts object containing the input time series}

\item{ts_output}{a vector or ts object (on the same time scale as ts_input) containing the target time series}

\item{iter}{number of iterations (maximum number of windows)}

\item{runs}{number of independent model runs}

\item{log}{whether a log-linear model should be used}

\item{parallel}{should the runs be computed in parallel? If FALSE, all runs are computed in serial. If TRUE, all runs are computed in parallel with a maximum number of cores. If a scalar is provided, the number of cores is set manually.}

\item{return}{either "best" (best model run is returned), or "all" (all model runs are returned)}

\item{param_selection}{either "max" (maximum number of windows), or "best_rmse", "best_aic", or "best_bic" to optimize RMSE, AIC, or BIC, respectively}
}
\description{
combines both training steps
}
\examples{
# train a model based on one year of observations
set.seed(42)
train(sampleWatershed$rain[1:365],
      sampleWatershed$gauge[1:365],
      iter = 2,
      runs = 1,
      parallel = FALSE)
}
